chore: apply session review/audit improvements — steps helper consolidation, knip 6.25, shellcheck baseline
steps/support/repo-fixtures.ts consolidates git/writeFileAt/audit-reader/
impl-plan fixtures behind both boundary-lane step files (quality-review
improvement 2, house pattern per steps/support/architecture-fixtures.ts);
knip bumped 6.24.0->6.25.0 (dev, minor, Low per audit triage); shellcheck
added to knip ignoreBinaries (system binary invoked by lint-staged — the
documented accepted-FP baseline). 53/53 cucumber green; knip clean.
