+++ knip.json
@@ -8,6 +8,7 @@
     "packages/cli/vitest.slow.config.ts"
   ],
   "ignoreBinaries": [
+    "shellcheck",
     "shfmt",
     "dot",
     "claude",
